requirements.txt
PyQt5
python-dotenv
together
requests
vtracer
Pillow
py2app
